+++ tests/framework/db/oci/QueryCacheTest.php
@@ -8,7 +8,7 @@ declare(strict_types=1);
  * @license https://www.yiiframework.com/license/
  */
 
-namespace yiiunit\framework\db\mssql;
+namespace yiiunit\framework\db\oci;
 
 use PHPUnit\Framework\Attributes\Group;
 use yii\caching\FileCache;
@@ -17,17 +17,17 @@ use yii\db\Query;
 use yiiunit\base\db\BaseDatabase;
 
 /**
- * Unit test for {@see \yii\db\Connection::cache()} with MSSQL driver.
+ * Unit test for {@see \yii\db\Connection::cache()} with Oracle driver.
  *
  * @author Wilmer Arambula <terabytesoftw@gmail.com>
  * @since 2.2
  */
 #[Group('db')]
-#[Group('mssql')]
+#[Group('oci')]
 #[Group('query-cache')]
 final class QueryCacheTest extends BaseDatabase
 {
-    protected $driverName = 'sqlsrv';
+    protected $driverName = 'oci';
 
     public function testQueryCacheFileCache(): void
     {
@@ -45,17 +45,17 @@ final class QueryCacheTest extends BaseDatabase
                 'type',
                 [
                     'int_col' => $key = 1,
-                    'char_col' => '',
+                    'char_col' => 'x',
                     'char_col2' => '6a3ce1a0bffe8eeb6fa986caf443e24c',
                     'float_col' => 0.0,
                     'blob_col' => 'a:1:{s:13:"template";s:1:"1";}',
-                    'bool_col' => true,
+                    'bool_col' => 1,
                 ],
             )
             ->execute();
 
         $value = static fn(Connection $db): bool|int|string|null => (new Query())
-            ->select(['blob_col'])
+            ->select(['char_col2'])
             ->from('type')
             ->where(['int_col' => $key])
             ->createCommand($db)
@@ -65,7 +65,7 @@ final class QueryCacheTest extends BaseDatabase
         $result = $db->cache($value);
 
         self::assertSame(
-            'a:1:{s:13:"template";s:1:"1";}',
+            '6a3ce1a0bffe8eeb6fa986caf443e24c',
             $result,
             'First query should return the correct scalar value.',
         );
@@ -74,7 +74,7 @@ final class QueryCacheTest extends BaseDatabase
         $result = $db->cache($value);
 
         self::assertSame(
-            'a:1:{s:13:"template";s:1:"1";}',
+            '6a3ce1a0bffe8eeb6fa986caf443e24c',
             $result,
             'Cached query should return the same scalar value.',
         );